updated 2pcf plotting functions
