+++ src/sp_validation/cosmo_val/psf_systematics.py
@@ -1113,231 +1113,6 @@ class PSFSystematicsMixin:
         self.print_done(f"Object-wise leakage coefficients plot saved to {out_path}")
 
     # --- utlility functions for plotting ---
-    def plot_2pcf_tomography(
-        self,
-        x_y_plot_function,
-        x_label,
-        y_label_plus,
-        y_label_minus,
-        tomo_bin_label_position,
-        extract_text_offset,
-        add_index_version_to_kwargs,
-        x_scale=None,
-        y_scale=None,
-        tomography=False,
-        versions=None,
-        colors=None,
-        savefig=None,
-        show=True,
-        close=True,
-        **kwargs,
-    ):
-        """
-        Standard plot function for 2-point correlation functions with tomographic bins.
-
-        Parameters
-        ----------
-        x_y_plot_function : callable
-            Function to plot the x and y data. Should accept axes for `+' and `-' components, version, tomo_bin_indices, and kwargs.
-        x_label : str
-            Label for the x-axis.
-        y_label_plus : str
-            Label for the y-axis of the `+' component.
-        y_label_minus : str
-            Label for the y-axis of the `-' component.
-        tomo_bin_label_position : tuple
-            Position to place the tomographic bin labels in axes coordinates (x, y).
-        extract_text_offset : bool
-            If True, extract the y-axis offset text and include it in the y-label.
-        add_index_version_to_kwargs : bool
-            If True, add the index of the version to the kwargs for plotting.
-        x_scale : str, optional
-            Scale for the x-axis ('linear', 'log', etc.). If None, default scale is used.
-        y_scale : str, optional
-            Scale for the y-axis ('linear', 'log', etc.). If None, default scale is used.
-        tomography : bool, optional
-            If True, plot the tomographic bins. Default is False.
-        versions : list, optional
-            List of versions to plot. If None, all versions are plotted.
-        colors : list, optional
-            List of colors for each version. If None, default colors are used.
-        savefig : str, optional
-            If provided, save the figure to this file.
-        show : bool, optional
-            If True, show the figure.
-        close : bool, optional
-            If True, close the figure after saving or showing.
-        kwargs : dict
-            Additional keyword arguments to pass to the plotting function.
-        """
-        versions = versions if versions is not None else self.versions
-        colors = (
-            colors
-            if colors is not None
-            else [self.cc[ver]["colour"] for ver in versions]
-        )
-
-        # First get the max number of tomo bins among the versions.
-        tomo_bins = self._get_tomo_bins_for_versions(versions, tomography=tomography)
-
-        max_key = max(tomo_bins, key=lambda k: len(tomo_bins[k]["ids"]))
-        n_tomo_bins_plot = len(tomo_bins[max_key]["ids"])
-        reference_tomo_bin_pairs = tomo_bins[max_key]["pairs"]
-
-        n_rows = n_tomo_bins_plot
-        n_cols = n_tomo_bins_plot + 2
-
-        # First start with the quantiles plot
-        fig, axs = plt.subplots(
-            n_rows,
-            n_cols,
-            figsize=(5 * n_cols, 5 * n_rows),
-            sharex=True,
-            sharey=True,
-            gridspec_kw={"wspace": 0, "hspace": 0},
-        )
-
-        for idx, (ver, color) in enumerate(zip(versions, colors)):
-            tomo_bin_pairs = tomo_bins[ver]["pairs"]
-
-            kwargs["color"] = color
-            if add_index_version_to_kwargs:
-                kwargs["idx"] = idx
-                kwargs["versions"] = versions
-
-            for tomo_bin_a, tomo_bin_b in tomo_bin_pairs:
-                # Plot the nsamples last samples
-                ax_plus = self._get_ax_plus(axs, tomo_bin_a, tomo_bin_b)
-                ax_minus = self._get_ax_minus(axs, tomo_bin_a, tomo_bin_b)
-
-                # Apply the x_y_plot_function to plot the data
-                x_y_plot_function(
-                    ax_plus, ax_minus, ver, tomo_bin_a, tomo_bin_b, **kwargs
-                )
-
-        # Draw to extract the y-axis text offset
-        fig.canvas.draw()
-
-        # Set the visibility to false where necessary
-        self._set_ax_visibility_to_false(axs, n_tomo_bins_plot)
-
-        # Set the labels and scales for the plots
-        for tomo_bin_a, tomo_bin_b in reference_tomo_bin_pairs:
-            ax_plus = self._get_ax_plus(axs, tomo_bin_a, tomo_bin_b)
-            ax_minus = self._get_ax_minus(axs, tomo_bin_a, tomo_bin_b)
-
-            ax_plus.tick_params(
-                axis="both",
-                which="both",
-                direction="in",
-                bottom=True,
-                top=False,
-                labelbottom=tomo_bin_b == 1 or tomo_bin_b == "all",
-                left=True,
-                right=False,
-                labelleft=tomo_bin_a == 1 or tomo_bin_a == "all",
-            )
-            if x_scale is not None:
-                ax_plus.set_xscale(x_scale)
-
-            ax_plus.text(
-                tomo_bin_label_position[0],
-                tomo_bin_label_position[1],
-                f"{tomo_bin_a}-{tomo_bin_b}",
-                transform=ax_plus.transAxes,
-                verticalalignment="top",
-                bbox=dict(
-                    boxstyle="square",
-                    facecolor="white",
-                    edgecolor="black",
-                    alpha=0.8,
-                ),
-            )
-            ax_plus.axhline(0, color="k", linestyle="--")
-
-            if y_scale is not None:
-                ax_plus.set_yscale(y_scale)
-            if tomo_bin_b == 1 or tomo_bin_b == "all":
-                ax_plus.set_xlabel(r"$\theta$ [arcmin]")
-            if tomo_bin_a == 1 or tomo_bin_a == "all":
-                text_offset = (
-                    ax_plus.yaxis.get_offset_text().get_text()
-                    if extract_text_offset
-                    else ""
-                )
-                ax_plus.set_ylabel(y_label_plus + text_offset)
-            ax_plus.yaxis.get_offset_text().set_visible(
-                False
-            )  # Hide the offset text for the plus ax
-
-            # Move the ticks to the right for the minus ax
-            ax_minus.yaxis.tick_right()
-            ax_minus.yaxis.set_label_position("right")
-            ax_minus.tick_params(
-                axis="both",
-                which="both",
-                direction="in",
-                bottom=True,
-                top=False,
-                labelbottom=tomo_bin_b == n_tomo_bins_plot or tomo_bin_b == "all",
-                left=False,
-                right=True,
-                labelleft=False,
-                labelright=tomo_bin_a == 1 or tomo_bin_a == "all",
-            )
-            if x_scale is not None:
-                ax_minus.set_xscale(x_scale)
-            ax_minus.text(
-                tomo_bin_label_position[0],
-                tomo_bin_label_position[1],
-                f"{tomo_bin_a}-{tomo_bin_b}",
-                transform=ax_minus.transAxes,
-                verticalalignment="top",
-                bbox=dict(
-                    boxstyle="square",
-                    facecolor="white",
-                    edgecolor="black",
-                    alpha=0.8,
-                ),
-            )
-            ax_minus.axhline(0, color="k", linestyle="--")
-            if y_scale is not None:
-                ax_minus.set_yscale(y_scale)
-            if tomo_bin_b == n_tomo_bins_plot or tomo_bin_b == "all":
-                ax_minus.set_xlabel(x_label)
-            if tomo_bin_a == 1 or tomo_bin_a == "all":
-                text_offset = (
-                    ax_minus.yaxis.get_offset_text().get_text()
-                    if extract_text_offset
-                    else ""
-                )
-                ax_minus.set_ylabel(y_label_minus + text_offset)
-            ax_minus.yaxis.get_offset_text().set_visible(
-                False
-            )  # Hide the offset text for the minus ax
-
-        # Build the legend
-        handles = []
-        for ver, color in zip(versions, colors):
-            label = self.cc[ver]["label"] if "label" in self.cc[ver] else ver
-            handles.append(plt.Line2D([0], [0], color=color, lw=2, label=label))
-        fig.legend(
-            handles=handles,
-            loc="upper center",
-            ncol=3,
-            frameon=False,
-            bbox_to_anchor=(0.5, 0.0),
-        )
-
-        if savefig is not None:
-            plt.savefig(savefig, dpi=300, bbox_inches="tight")
-            self.print_done(f"Plot saved to {os.path.abspath(savefig)}")
-
-        if show:
-            plt.show()
-
-        if close:
-            plt.close()
 
     def _xi_psf_sys_sample_x_y_plot_function(
         self,
